+++ applications/amathea/src/features/modal/Modal.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense, lazy, useCallback, useMemo } from 'react'
-import { ModalConfigOptions, ModalTypes } from 'types'
 import GFWModal from '@globalfishingwatch/ui-components/dist/modal'
 import Spinner from '@globalfishingwatch/ui-components/dist/spinner'
+import { ModalConfigOptions, ModalTypes } from 'types'
 import { useCurrentWorkspaceConnect } from 'features/workspaces/workspaces.hook'
 import { useDraftDataviewConnect } from 'features/dataviews/dataviews.hook'
 import { useDraftDatasetConnect } from 'features/datasets/datasets.hook'