+++ applications/temporalgrid-demo/src/map.js
@@ -1,56 +1,45 @@
-import React, {useState, useRef, useEffect, useMemo} from 'react';
-import MapGL from '@globalfishingwatch/react-map-gl';
+import React, { useState, useRef, useEffect, useMemo } from 'react'
+import MapGL from '@globalfishingwatch/react-map-gl'
 
 const TEST_GEO_JSON = {
-  "type": "FeatureCollection",
-  "features": [
+  type: 'FeatureCollection',
+  features: [
     {
-      "type": "Feature",
-      "properties": {
-        value: "some eez"
+      type: 'Feature',
+      properties: {
+        value: 'some eez',
       },
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
+      geometry: {
+        type: 'Polygon',
+        coordinates: [
           [
-            [
-              -12.65625,
-              41.244772343082076
-            ],
-            [
-              -3.69140625,
-              41.11246878918088
-            ],
-            [
-              -5.2734375,
-              44.715513732021336
-            ],
-            [
-              -7.03125,
-              46.92025531537451
-            ],
-            [
-              -10.8984375,
-              45.706179285330855
-            ],
-            [
-              -12.65625,
-              41.244772343082076
-            ]
-          ]
-        ]
-      }
-    }
-  ]
+            [-12.65625, 41.244772343082076],
+            [-3.69140625, 41.11246878918088],
+            [-5.2734375, 44.715513732021336],
+            [-7.03125, 46.92025531537451],
+            [-10.8984375, 45.706179285330855],
+            [-12.65625, 41.244772343082076],
+          ],
+        ],
+      },
+    },
+  ],
 }
 
-const Map = React.memo(function Map({ style, onMapClick, onMapHover, onSetMapRef, children, tiles }) {
+const Map = React.memo(function Map({
+  style,
+  onMapClick,
+  onMapHover,
+  onSetMapRef,
+  children,
+  tiles,
+}) {
   // console.log(tiles)
   const [viewport, setViewport] = useState({
     longitude: 13,
     latitude: -30,
-    zoom: 1.58
-  });
+    zoom: 1.58,
+  })
 
   const mapRef = useRef(null)
 
@@ -58,7 +47,7 @@ const Map = React.memo(function Map({ style, onMapClick, onMapHover, onSetMapRef
     if (onSetMapRef && mapRef.current) {
       onSetMapRef(mapRef.current.getMap())
     }
-  }, [mapRef]);
+  }, [mapRef])
 
   const customStyle = useMemo(() => {
     if (!style) return null
@@ -66,11 +55,11 @@ const Map = React.memo(function Map({ style, onMapClick, onMapHover, onSetMapRef
       ...style,
       sources: {
         ...style.sources,
-        'test': {
+        test: {
           type: 'geojson',
           data: TEST_GEO_JSON,
           generateId: true,
-        }
+        },
       },
       layers: [
         ...style.layers,
@@ -85,10 +74,10 @@ const Map = React.memo(function Map({ style, onMapClick, onMapHover, onSetMapRef
               ['boolean', ['feature-state', 'hover'], false],
               '#ffffff',
               '#000000',
-            ]
-          }
-        }
-      ]
+            ],
+          },
+        },
+      ],
     }
   }, [style])
 
@@ -99,19 +88,21 @@ const Map = React.memo(function Map({ style, onMapClick, onMapHover, onSetMapRef
     return null
   }, [customStyle])
 
-  return <MapGL
-    {...viewport}
-    ref={mapRef}
-    width="100%"
-    height="100%"
-    mapStyle={customStyle}
-    onViewportChange={setViewport}
-    onClick={onMapClick}
-    onHover={onMapHover}
-    interactiveLayerIds={customInteractiveLayerIds}
-  >
-    {children}
-  </MapGL>
+  return (
+    <MapGL
+      {...viewport}
+      ref={mapRef}
+      width="100%"
+      height="100%"
+      mapStyle={customStyle}
+      onViewportChange={setViewport}
+      onClick={onMapClick}
+      onHover={onMapHover}
+      interactiveLayerIds={customInteractiveLayerIds}
+    >
+      {children}
+    </MapGL>
+  )
 })
 
 export default Map