+++ packages/layer-composer/src/generators/heatmap/heatmap-animated.ts
@@ -6,12 +6,8 @@ import {
   HeatmapAnimatedMode,
 } from '../types'
 import { memoizeByLayerId, memoizeCache } from '../../utils'
-import {
-  API_TILES_URL,
-  API_ENDPOINTS,
-  HEATMAP_DEFAULT_MAX_ZOOM,
-  HEATMAP_MODE_COMBINATION,
-} from './config'
+import { API_GATEWAY } from '../../layer-composer'
+import { API_ENDPOINTS, HEATMAP_DEFAULT_MAX_ZOOM, HEATMAP_MODE_COMBINATION } from './config'
 import { TimeChunk, TimeChunks, getActiveTimeChunks, getDelta } from './util/time-chunks'
 import { getSublayersBreaks } from './util/get-legends'
 import getGriddedLayers from './modes/gridded'
@@ -27,7 +23,6 @@ const DEFAULT_CONFIG: Partial<HeatmapAnimatedGeneratorConfig> = {
   tilesetsStart: '2012-01-01T00:00:00.000Z',
   tilesetsEnd: new Date().toISOString(),
   maxZoom: HEATMAP_DEFAULT_MAX_ZOOM,
-  tilesAPI: API_TILES_URL,
   interactive: true,
 }
 
@@ -52,7 +47,7 @@ class HeatmapAnimatedGenerator {
     const datasets = config.sublayers.map((sublayer) => sublayer.datasets.join(','))
     const filters = config.sublayers.map((sublayer) => sublayer.filter || '')
 
-    const tilesUrl = `${config.tilesAPI}/${API_ENDPOINTS.tiles}`
+    const tilesUrl = `${config.tilesAPI || `${API_GATEWAY}/v1`}/${API_ENDPOINTS.tiles}`
 
     const delta = getDelta(timeChunks.activeStart, timeChunks.activeEnd, timeChunks.interval)
     const breaks = getSublayersBreaks(config, timeChunks.deltaInDays)